appengine/server/requirements.txt
google-endpoints==2.4.5
google-endpoints-api-management==1.3.0
googleapis_common_protos==1.2.0
gcloud==0.18.1
google-api-python-client==1.6.4
oauth2client==3.0.0
